requirements.txt
atlassian-python-api
semantic_version
datetime